requirements.txt
mintotp==0.3.0
python-dotenv==1.0.1
Requests==2.32.3
selenium==4.25.0
